features:
		- integration test framework
		- x86 code coverage support
		- x86 architecture layer
		- x86 uart driver
		- refactor lib initialisation and add support for
		  __[pre]init_array

	fixes:
		- kernel timer time conversion
		- build system code coverage builds
		- memory alignment for linker arrays
		- docu build targets now use cmd_run_script
		- invalid handling of test-case return value
		- type cast warning in dup()
		- missing initialisation in init shell
		- null-pointer for '%s' specifier in vfprintf()
		- init shell thread test

	miscellaneous:
		- move functions for interrupt registration from arch to kernel
		- update kernel startup output
		- flush kernel log buffer once a log device is available
		- rework kernel log buffer overflow handling
		- move unit tests into dedicated directory
		- speedup docu builds and cleanup their output
		- remove thread pointer from kpanic calls
		- remove p_err parameter from exec_init_call()
		- utility functions
			- kpmalloc()
			- stack_pop()

		- fix includes for sys/list
		- fix dependencies
			- scheduler syscalls -> signal syscalls
			- init shell -> init shell history
			- remove avr core.o from libsys

		- always include all host tests
		- usage of CPP_ERROR instead of GCC_ERROR, renaming ERROR to ASSERT
		- remove avr timer header since its not needed
		- unify line break usage when calling kpanic()
		- fix source file section order, "external variables" in particular

	issue: 33

+++ include/sys/time.h
@@ -25,8 +25,8 @@
 /* types */
 typedef struct{
 	uint32_t s;
-	size_t ms,
-		   us;
+	uint16_t ms,
+			 us;
 } time_t;
